django_certiffy_project/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
crispy-bootstrap5==2023.10
cryptography==41.0.7
Django==4.2.4
django-crispy-forms==2.1
django-crontab==0.7.1
django-extensions==3.2.3
django-multi-email-field==0.7.0
django-mysql==4.12.0
dnspython==2.4.2
gunicorn==21.2.0
idna==3.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.1
pandas==2.2.0
pathspec==0.11.2
platformdirs==3.10.0
plotly==5.18.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.4
